Updated imageMaker.pymaidenhead_C.py templates.py .
