+++ imageMaker.py
@@ -5,7 +5,7 @@ import maidenhead_C
 MH = maidenhead_C.maidenhead()
 
 
-print(f"""{MH.latLonToMaidenhead(lat_= 40.687088, lon_=-111.927630, chrs_=12)}""")
+print(f"""{MH.latLonToMaidenhead(lat_=  0, lon_=-180, chrs_=12)}""")
 
 
 #
