+++ maidenhead_C.py
@@ -3,100 +3,6 @@
 from CF.CONST_D import _00_VALS as CFV
 
 
-def DOCUMENT_HEAD_STR(*,
-		abstractView=None,
-		address_=None,
-		addressDetails=None,
-		author_=None,
-		description=None,
-		extendedData=None,
-		id_=None,
-		isopen_=None,
-		link_=None,
-		name_=None,
-		phoneNumber=None,
-		region=None,
-		snippet=None,
-		styleSelector=None,
-		styleUrl=None,
-		timePrimitive=None,
-		visibility_=None,
-):
-	# fold here ⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1
-	_strToRtn_ = ""
-	# * 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱ 1⟱
-	if (
-			(id_ is None)
-	):
-		_strToRtn_ += f""""""
-	# * ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1 ⟰1
-
-	f"""
-<Document id="ID">
-	<!-- inherited from Feature element -->
-	<name>...</name>                      <!-- string -->
-	<visibility>1</visibility>            <!-- boolean -->
-	<open>0</open>                        <!-- boolean -->
-	<atom:author>...<atom:author>         <!-- xmlns:atom -->
-	<atom:link href=" "/>                 <!-- xmlns:atom -->
-	<address>...</address>                <!-- string -->
-	<xal:AddressDetails>...</xal:AddressDetails>  <!-- xmlns:xal -->
-	<phoneNumber>...</phoneNumber>        <!-- string -->
-	<Snippet maxLines="2">...</Snippet>   <!-- string -->
-	<description>...</description>        <!-- string -->
-	<AbstractView>...</AbstractView>      <!-- Camera or LookAt -->
-	<TimePrimitive>...</TimePrimitive>
-	<styleUrl>...</styleUrl>              <!-- anyURI -->
-	<StyleSelector>...</StyleSelector>
-	<Region>...</Region>
-	<Metadata>...</Metadata>              <!-- deprecated in KML 2.2 -->
-	<ExtendedData>...</ExtendedData>      <!-- new in KML 2.2 -->
-
-	<!-- specific to Document -->
-	<!-- 0 or more Schema elements -->
-	<!-- 0 or more Feature elements -->
-</Document>
-"""
-	# fold here ⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1
-
-GRID_POLY_STR = lambda __NAME__, __VISIBILITY__, __STYLE_ID__, __WEST__, __SOUTH__, __NORTH__, __EAST__, __ALT__: f"""
-	<Placemark>
-		<name>{__NAME__}</name>
-		<visibility>{__VISIBILITY__}</visibility>
-		<styleUrl>{__STYLE_ID__}</styleUrl>
-		<Polygon>
-			<tessellate>1</tessellate>
-			<outerBoundaryIs>
-				<LinearRing>
-					<coordinates>
-						{__WEST__},{__SOUTH__},0 {__WEST__},{__NORTH__},0 {__EAST__},{__NORTH__},0 {__EAST__},{__SOUTH__},0 {__WEST__},{__SOUTH__},0
-					</coordinates>
-				</LinearRing>
-			</outerBoundaryIs>
-		</Polygon>
-	</Placemark>
-"""
-
-STYLE_STR = lambda __STYLE_ID__, __ICON_URI__, __HOTSPOT_X__, __HOTSPOT_Y__, __HOTSPOT_X_UNIT__, __HOTSPOT_Y_UNIT__, __LINE_COLOR__, __POLY_COLOR__, __COLOR_MODE__: f"""
-			<Style id="{__STYLE_ID__}">
-				<IconStyle>
-					<scale>1.1</scale>
-					<Icon>
-						<href>{__ICON_URI__}</href>
-					</Icon>
-					<hotSpot x="{__HOTSPOT_X__}" y="{__HOTSPOT_Y__}" xunits="{__HOTSPOT_X_UNIT__}" yunits="{__HOTSPOT_Y_UNIT__}"/>
-				</IconStyle>
-				<LineStyle>
-					<color>{__LINE_COLOR__}</color>
-				</LineStyle>
-				<PolyStyle>
-					<color>{__POLY_COLOR__}</color>
-					<colorMode>{__COLOR_MODE__}</colorMode>
-				</PolyStyle>
-			</Style>
-		"""
-
-
 class maidenhead(object):
 	# fold here ⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1⟱1
 
@@ -193,9 +99,21 @@ class maidenhead(object):
 	# * #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*
 	def latLonToMaidenhead(self, *,
 			chrs_,  # how many characters 2, 4, 6, 8, 10, 12) of maidenhead to return
-			lat_,
-			lon_,
+			lat_,  # -90 to 90 D.ddddddddddddd
+			lon_,  # -180 to 180 D.ddddddddddddd
 	):
+		"""
+Input is:
+		chrs_		How many characters of maidenhead to return, must be even from 2 to 12.
+		lat_=  40.687216, lon_=-111.927653, chrs_=12
+				 2 = DN
+				 4 = DN40
+				 6 = DN40aq
+				 8 = DN40aq84
+				10 = DN40aq84qw
+				12 = DN40aq84qw33
+Output is a Maidenhead string with specified character length.
+		"""
 		# fold here ⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2
 		_lat_ = lat_ + 90
 		_lon_ = lon_ + 180
@@ -307,6 +225,26 @@ class maidenhead(object):
 		# fold here ⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2
 	# * #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*
 	# end of latLonToMaidenhead
+	# * #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*
+
+	# * #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*
+	# start of makeKMLGrids
+	# * #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*
+	def makeKMLGrids(self *,
+			depth_=6,
+			lat_=None,
+			lon_=None,
+			maidenheadGrid_=None,
+	):
+		"""
+Input is the first found of lat/lon D.ddddddddddddd or maidenhead string.
+
+		"""
+		# fold here ⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2⟱2
+		#
+		# fold here ⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2⟰2
+	# * #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*
+	# end of makeKMLGrids
 	# * #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*
 
 	# fold here ⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1⟰1
